Added casts and do-nothing accesses of unused variables to remove all warnings from the build (and we're using effc++ !!!
git-svn-id: svn://svn.code.sf.net/p/xlslib/code/trunk@137 126565e4-af4f-46c5-a216-0262b1fa06ba

+++ xlslib/src/xlslib/globalrec.cpp
@@ -868,7 +868,7 @@ void  CGlobalRecords::wide2str16(const ustring& str1, u16string& str2)
 	if(resultSize == (size_t)-1) {
 		str2 = convFail;
 	} else {
-		str2.assign(origOutbuf, outbuf - origOutbuf);
+		str2.assign(origOutbuf, (size_t)(outbuf - origOutbuf));
 	}
 	free((void *)origOutbuf);
 }
